% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsExample.R
\docType{data}
\name{grsExample}
\alias{grsExample}
\title{Filtered Mock Epigenomics Data Set}
\format{An object of class \code{GenomicRatioSet} with 400 rows and 210 columns.}
\usage{
grsExample
}
\value{
A \code{GenomicRatiSet} containing pre-filtered CpG sites, ready to
        be used with the core function of the \code{methyvim} package.
}
\description{
A data set containing data that is similar in structure to that typically
found in epigenomics studies.
}
\details{
\describe{
  This data set was simulated based on several key characteristics of data
  sets available to the package authors through collaborations in studies of
  differential methylation. While the motivating data sets were collected on
  the Infinium MethylationEPIC BeadChip (Illumina) platform, the data has
  been reduced for convenience purposes. We emphasize that this is not real
  data and is not in any way suitable for scientific investigation. Please
  note that THIS IS NOT REAL DATA.
}
}
\keyword{datasets}
