% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsExample.R
\docType{data}
\name{grsExample}
\alias{grsExample}
\title{Filtered Mock Epigenomics Data Set}
\format{An object of class \code{GenomicRatioSet} with 400 rows and 210 columns.}
\usage{
grsExample
}
\value{
A \code{GenomicRatioSet} containing pre-filtered CpG sites, ready to
        be used with the core function of the \code{methyvim} package.
}
\description{
A data set containing data that is similar in structure to that typically
found in epigenomics studies.
}
\details{
\describe{
  This data set was simulated based on several key characteristics of data
  sets available to the package authors through collaborations in studies of
  differential methylation. It is based on data collected using the 850K
  (EPIC) Infinium arrays produced by Illumina, from samples collected on
  human subjects in studies of environmental epidemiology. The structure of
  the original data set (e.g., notable differences between the exposure and
  treatment conditions) were used to guide the simulation, and data was only
  produced for an arbitrary subset of sites. The phenotype-level exposure and
  outcome data (accessible via \code{colData}) were derived from the original
  data. The size of the subset was chosen such that the data package would be
  compliant with the standards outlined by Bioconductor. We emphasize that
  this is not real data and is not in any way suitable for scientific
  investigation. Please note that THIS IS NOT REAL DATA.
}
}
\keyword{datasets}
